+++ .meteor/versions
@@ -1,5 +1,5 @@
 accounts-base@2.2.2
-accounts-oauth@1.4.0
+accounts-oauth@1.4.1
 accounts-password@2.3.1
 allow-deny@1.1.1
 autoupdate@1.8.0
@@ -60,7 +60,7 @@ mongo-decimal@0.1.2
 mongo-dev-server@1.1.0
 mongo-id@1.0.8
 npm-mongo@4.3.1
-oauth@2.1.1
+oauth@2.1.2
 oauth2@1.3.1
 observe-sequence@1.0.19
 ordered-dict@1.1.0
