+++ imports/startup/server/index.js
@@ -24,5 +24,4 @@ import '../../api/extensions/methods.js'
 import '../../api/extensions/server/publications'
 import '../../api/customfields/server/methods.js'
 import '../../api/customfields/server/publications'
-import '../../api/openid/server/methods'
 import './startup.js'
