+++ imports/startup/server/startup.js
@@ -1,6 +1,7 @@
 import { AccountsAnonymous } from 'meteor/faburem:accounts-anonymous'
 import Extensions from '../../api/extensions/extensions.js'
 import { defaultSettings, Globalsettings } from '../../api/globalsettings/globalsettings.js'
+import { getGlobalSetting } from '../../utils/frontend_helpers'
 
 Meteor.startup(() => {
   AccountsAnonymous.init()
@@ -16,6 +17,11 @@ Meteor.startup(() => {
   if (Meteor.settings.enableAnonymousLogins) {
     // eslint-disable-next-line i18next/no-literal-string
     Globalsettings.update({ name: 'enableAnonymousLogins' }, { $set: { value: Meteor.settings.disablePublic === 'true' } })
+  }
+  if (getGlobalSetting('enableOpenIDConnect')) {
+    import('../../utils/oidc_server').then((Oidc) => {
+      Oidc.registerOidc()
+    });
   }
   for (const extension of Extensions.find({})) {
     if (extension.isActive) {
